+++ wsl-provision.sh
@@ -36,23 +36,6 @@ if check_command sudo ; then
   sudo apt-get install -y wget cups thunar
 fi
 
-# Offer to add user to lpadmin
-if groups | grep -F "lpadmin" >/dev/null 2>&1 ; then
-  echo "Local Printer Admin:"
-  echo_continue "OK (lpadmin)"
-else
-  if check_command sudo ; then
-    if check_yesno "Add $USER to group lpadmin for CUPS printer admin?" ; then
-      sudo gpasswd -a "$USER" "lpadmin"
-        echo_continue "Done"
-    else
-      echo_continue "Skipped"
-    fi
-  else
-    check_continue "Add user to group lpadmin to access serial port."
-  fi
-fi
-
 # Start Metarace AutoInstaller
 sh /mnt/c/temp/metarace-install.sh